SDFsd

+++ train_PCNN_GAN.py
@@ -21,10 +21,10 @@ if __name__ == "__main__":
 
     iters_per_epoch = (data_train.num_imgs - test_size) // batch_size
 
-    generator = PartialConvNet().float()
+    generator = PartialConvNet().half()
     generator = torch.nn.DataParallel(generator)
     generator = generator.to(device)
-    discriminator = Discriminator().float()
+    discriminator = Discriminator().half()
     discriminator = torch.nn.DataParallel(discriminator)
     discriminator = discriminator.to(device)
     print("Loaded model to device...")
@@ -68,7 +68,7 @@ if __name__ == "__main__":
         monitor_dis_loss = 0
         for i in range(1, iters_per_epoch+1):
             print(i)
-            image, mask, gt = [x.float().to(device) for x in next(iterator_train)]
+            image, mask, gt = [x.half().to(device) for x in next(iterator_train)]
 
             pred_img = generator(image, mask)
             comp_img = (1 - mask) * gt + mask * pred_img
@@ -98,7 +98,7 @@ if __name__ == "__main__":
                 test_losses = list()
                 with torch.no_grad():
                     for k in range(test_size):
-                        image, mask, ground_truth = [x.float().to(device) for x in data_train[data_train.num_imgs - test_size + k]]
+                        image, mask, ground_truth = [x.half().to(device) for x in data_train[data_train.num_imgs - test_size + k]]
                         image, mask, ground_truth = image[None, :, :, :], mask[None, :, :, :], ground_truth[None, :, :, :]
 
                         pred_img = generator(image, mask)
